创建XDCacheManager用于管理缓存，可以配置用于缓存的磁盘容量和缓存将原本处理缓存的功能从XDNetworking+cache转移到XDCacheManager；优化LRU算法，遍历查找改为逆序查找，因为经常使用的缓存一般位于数组的尾部，在查找缓存的时候可以减少遍历的次数；为每个缓存配置一个缓存时间，原本是想抽象成对象，但考虑到对象转化存储需要时间，而且只有两个属性，所以就直接用字典

+++ XDNetworking/XDNetworking/XDNetworking/Cache/XDDistCache.h
@@ -42,7 +42,7 @@
  *
  *  @return 文件总大小
  */
-+ (unsigned long long)dataSizeInDir:(NSString *)directory;
++ (NSUInteger)dataSizeInDir:(NSString *)directory;
 
 /**
  *  清理目录中的文件
